+++ src/components/SignUp/SignUp.js
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import { useState } from "react";
 
-import UserService from "services/userService";
+import UserService from "services/UserService";
 import Navbar from "components/Navbar/Navbar";
 import FormInput from "components/FormInput/FormInput";
 import FormSelect from "components/FormSelect/FormSelect";
